pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["ssyucfa <patin123321@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
PyMySQL = "^1.0.2"
mysqlclient = "^2.1.0"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
